requirements.txt
Pillow==4.1.0
numpy==1.12.1
opencv-contrib-python==3.2.0.7
fire==0.1.0
Flask==0.12.1